consolidate challenge code in _shared
