+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/_shared/__init__.py
@@ -2,7 +2,12 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-__all__ = []  # type: ignore
+from .challenge_auth_policy import ChallengeAuthPolicy
+from .http_challenge import HttpChallenge
+from . import http_challenge_cache as HttpChallengeCache
+
+__all__ = ["ChallengeAuthPolicy", "HttpChallenge", "HttpChallengeCache"]
+
 try:
     from .async_challenge_auth_policy import AsyncChallengeAuthPolicy
