+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/_shared/http_challenge_cache.py
@@ -7,7 +7,6 @@ import threading
 try:
     import urllib.parse as parse
 except ImportError:
-    # pylint: disable=import-error
     import urlparse as parse  # type: ignore
 
 try:
